Day 22
- Fixed typo in stats footer
- Started working on tournament browser
- Changed version name from "Alpha" to "Pre-Alpha" because let's be
real, it's too small to be called alpha
- Fixed wrong hover colors on navbar links
- Changed slogan to "Circles frontpage: the Pulse of osu! competitive
play" because this sound better and I actually like it
- Added spoiler blockers

I'm back (hopefully). This is my first time working on this project in
two months. For the 0 people reading this I'll try to explain why. Main
reason: idk actually. There's currently 75 folders in my old projects
folder, and 5 more on a new drive. Out of these 80 projects I only
finished 3. Only because I had no choice. It's really hard for me to
finish anything. There's usually me working day and night for a 3 weeks
on project just for it to be forgoten on week 4. I don't why this is
happening. On week 4 I open VScodium and do nothing for 30 minutes
before I close it and forget. Even if I still really want to finish
(like in case of cfp) I just can't sit and write code with the same
passion as on my first week.
I came back because 1. I really want to be known in the community, and 2
I was thinking a lot about finishing it every day, which gave me a lot
of extra motivation.
I really hope that it will be used and loved by the community.
Thank you for reading